% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{add.data}
\alias{add.data}
\title{Augment Data Variables}
\usage{
add.data(new.path, old.data, scaled = F, mods, filts = NULL)
}
\arguments{
\item{new.path}{Path to Stata .dta file containing new variables}

\item{old.data}{Tibble of old study data to merge with}

\item{scaled}{Boolean indicating whether new data should be scaled}

\item{mods}{List of paths to formatted csv files creating new adjustment
variables. Format of csv is detailed below. This parameter is
required. See \code{\link[=get.adjust]{get.adjust()}} for formatting.}

\item{filts}{List of path (singular) to formatted csv of the data filters to
be applied to the dataset. See \code{\link[=get.adjust]{get.adjust()}} for formatting.}
}
\value{
Tibble with new variables joined in.
}
\description{
This function serves to add variables to the master tibble dataframe.
Variables are joined in on the id column.
}
\details{
See \code{\link[=get.adjust]{get.adjust()}} for csv formatting details.
}
\examples{
\dontrun{
fifth.visit.adtl <- add.data(new.path = 'data/ARIC_proteomics_addlvar_01062021.dta',
                             old.data = haven::read_dta('data/ARICmaster_121820.dta'),
                             mods = list('data/visit-five/adtlvars.csv'))
}
}
