% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-array.R
\name{cox.arry}
\alias{cox.arry}
\title{Cox Array Main Function}
\usage{
cox.arry(proteins, data, time, outc, adjust, .src)
}
\arguments{
\item{proteins}{This is the vector of all proteins to be considered for
Uniprotein, Adjusted Cox Regressions.}

\item{data}{This is the tibble dataframe with the covariates for regression}

\item{time}{This is a character vector of the Survival time to event Outcome.}

\item{outc}{This is a character vector of the Survival event indicator
Outcome.}

\item{adjust}{This is a vector of all the adjustment variables.}

\item{.src}{This is an indicator of whether the models are the derivation or
validation models. It is 1 when the data is the derivation dataset, and 2
when it is the validation dataset.}
}
\value{
a Tibble containing the results of the multiple regression models.
}
\description{
This is the main access point for using parallel processing to complete
numerous Cox regressions simultaneously.
}
\details{
This function utilizes the \code{\link[furrr:future_map]{furrr::future_map()}} function to complete many
regressions quickly. Each protein is regressed individually against the time
and outcome variable specified, while adjusted for the provided variables.

The .src input is specified to distinguish data between derivation data and
validation data. Derivation data should be \code{.src = 1} while validation
data should be \code{.src = 2}.
}
\examples{
\dontrun{
  data <- haven::read_dta('~/proteomics/studydata.dta') \%>\% ...
      (merge study data with SomaData, see [get.visit()] for specifics)

  univariate.results <- cox.arry(labels$term, data,
                                 time = "fuptime",
                                 outc = "hfdiag",
                                 adjust = adjust,
                                 .src = 1)

 validation.results <- cox.arry(labels$term, data.validation,
                                 time = "fuptime",
                                 outc = "hfdiag",
                                 adjust = adjust,
                                 .src = 2)

}
}
