% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-array.R
\name{cox.modl}
\alias{cox.modl}
\title{Cox Model Wrapper Function}
\usage{
cox.modl(
  .data,
  time = "fuptime",
  outcome = "hfdiag",
  protein,
  .adjust = adjust,
  extend = 100
)
}
\arguments{
\item{.data}{The data tibble containing the variables to be regressed.}

\item{time}{A character vector of the outcome variable time to event}

\item{outcome}{A character vector of the outcome variable event indicator}

\item{protein}{A character vector indicating which protein is being
regressed.}

\item{.adjust}{A vector of all adjustment variables for this model.}

\item{extend}{Number of asterisks to add to Pvalue table indicator.}
}
\value{
A list with three components: \describe{ \item{hazr}{The Hazard Ratio
of the Protein} \item{cint}{A Vector of length 2 with Lower and Upper
Confidence Interval} \item{pval}{The P value of the model covariate}
\item{desc}{A string that combines the relevant information (such as hazard
ratio and confidence interval) needed to display in a table.}}
}
\description{
This function serves to wrap the survival coxph function for easier usage in
evaluating many cox models simultaneously. This function is accessed by
\code{\link[=cox.arry]{cox.arry()}} in order to parallelize the univariate comparisons of every
protein against the outcome.
}
\details{
This function returns the effect estimate for the "protein" of interest
adjusted. The function also creates a string descriptor to print output as
table for \code{.csv} files. The significance of the p value is indicated by
each marker denoting a power of ten smaller.
}
\examples{
\dontrun{

  data <- haven::read_dta('~/proteomics/studydata.dta')

  cox <- cox.modl(data, time = 'adjudhfdate', outcome = 'adjudhf_bwh',
                  protein = 'SeqId_7655_11', .adjust = adjust)

}


}
