% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic-net.R
\name{elastic.net}
\alias{elastic.net}
\title{Main Elastic Net Wrapper Function}
\usage{
elastic.net(data, terms, adjust, time, outc, .alpha = 1)
}
\arguments{
\item{data}{The tibble containing the merged Protein and Study Data.}

\item{terms}{The terms to consider in the LASSO Regression.}

\item{adjust}{The vector of Adjustment variables to force}

\item{time}{A character vector of the Survival time to event outcome
variable.}

\item{outc}{A character vector of the Survival event indicator outcome
variable.}

\item{.alpha}{The optional alpha parameter to change LASSO to Elastic Net.}
}
\value{
\describe{
\item{model}{The returned cv.glmnet LASSO Elastic Net Model.}
\item{coefs}{The coefficients of the Non-zero covariates retained.}
\item{terms}{The Final retained Proteins from this model.}
}
}
\description{
This function is a wrapper function to the \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} function to
conduct LASSO / ElasticNet regularized regression. The regression optimizes
the C statistic, forces the adjustment variables, and runs the algorithm
using parallel processing.
}
\examples{
\dontrun{

 fifth.lasso <- elastic.net(fifth.visit,
                            bonferroni$kept,
                            adjust, time, outc)

}
}
