% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{get.data}
\alias{get.data}
\title{Data Joining, Imputation, and Scaling}
\usage{
get.data(master, soma.data, adjust)
}
\arguments{
\item{master}{Master study data tibble from \code{\link[haven:read_dta]{haven::read_dta()}}}

\item{soma.data}{Parsed SomaLogic proteomic data tibble from \code{\link[=get.soma]{get.soma()}}}

\item{adjust}{List of Adjustment variables. Must match \code{adjusted.csv}
files, but should not include id variable.}
}
\value{
Tibble of joined and cleaned proteomics and study data.
}
\description{
Data Joining, Imputation, and Scaling
}
\examples{
\dontrun{
labels <- get.labels('data/proteinMapping.csv', sid,
                     uniprot.full.name,
                     entrezgenesymbol,
                     flag2 == 0)

soma.data.fifth <- get.soma('data/somaV5.txt', labels = labels)

fifth.visit.study <-
  get.adjust(stata = haven::read_dta('data/ARICmaster_121820.dta'),
             mods  = list('data/visit-five/outcomes.csv',
                          'data/visit-five/adjusted.csv'),
             filts = list('data/visit-five/filters.csv'))

adj <- readr::read_csv('data/visit-five/adjusted.csv')[[1]][-1]

visit.data <- get.data(fifth.visit.study, soma.data.fifth, adj)
}
}
