% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{get.filtered}
\alias{get.filtered}
\title{Filter Data given series of filter commands}
\usage{
get.filtered(data, filts)
}
\arguments{
\item{data}{The Tibble of Data from Study to be Filtered}

\item{filts}{The list of file path(s) to \code{filters.csv}. Most likely a
list of length 1.}
}
\value{
returns a filtered tibble of study data. Does not do any error checking.
}
\description{
This function is a helper function to take a series of filter instructions
(from the \code{filters.csv} file) and uses them to filter a tibble of study
data.
}
\details{
This command writes an intermediate csv file to the working directory, named
\code{filres.csv}. This intermediate file stores the data about how many
subjects were filtered by each condition. This data is later used to create
consort diagrams.
}
\examples{
\dontrun{

filts <- '~/proteomics/filters.csv'
data  <- haven::read_dta('~/proteomics/studydata.dta')

filtered.data <- get.filtered(data, filts)

OR

data <- haven::read_dta('~/proteomics/studydata.dta') \%>\%
    get.filtered('~/proteomics/filters.csv')

}

}
