% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{get.labels}
\alias{get.labels}
\title{Get Protein Names and Labels from SeqId Identifiers}
\usage{
get.labels(path, term, name, gene, flag)
}
\arguments{
\item{path}{File Path to SomaLogic Data Dictionary}

\item{term}{Column Variable ID for the SeqId term}

\item{name}{Column Variable ID for the UniProt Full Name Column}

\item{gene}{Column Variable ID for the Entrez Gene Symbol Column}

\item{flag}{Filtering Function for Proteins that Pass QC}
}
\value{
Tibble with columns \code{term} for SeqId, \code{name} for
UniProt Name, and \code{gene} for Entrez Gene Symbol. Data Frame rownames
are set as the SeqId identifiers.
}
\description{
This is a function to generate a dictionary to translate from SeqId to
Protein Name and Entrez Gene Symbol. Output from this function is used in
creating tables, figures, and labels for SeqIds. This function also requires
an input for data cleaning flags. Proteins that are flagged will not be
included in the final label frame.
}
\details{
The term, name, gene, and flag functions must be provided as-is (not
as strings). This is a quirk of using tidyverse and dplyr.
}
\examples{
\dontrun{
labels <- get.labels('data/proteinMapping.csv', sid,
                     uniprot.full.name,
                     entrezgenesymbol,
                     flag2 == 0)
}

}
