% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{get.scaled}
\alias{get.scaled}
\title{Scale Data}
\usage{
get.scaled(data, adjust, proteins = NULL)
}
\arguments{
\item{data}{Tibble of Data to be scaled}

\item{adjust}{List of NUMERIC ADJUSTMENT VARS ONLY -- output of
\code{\link[=numeric.id]{numeric.id()}}.}

\item{proteins}{List of Proteins to Scale}
}
\value{
Tibble where all numeric adjustment variables and protein variables
are scaled to mean 0 and sd 1.
}
\description{
This function scales data after tables are made. Scales numeric adjustment
and proteomic variables. Done separately to maintain scale and values for
table 1 and figure generation. Numeric adjustment variables must be
specified. See \code{\link[=numeric.id]{numeric.id()}} for more details.
}
\examples{
\dontrun{
labels <- get.labels('data/proteinMapping.csv', sid,
                     uniprot.full.name,
                     entrezgenesymbol,
                     flag2 == 0)

soma.data.fifth <- get.soma('data/somaV5.txt', labels)

fifth.visit.study <-
  get.adjust(stata = haven::read_dta('data/ARICmaster_121820.dta'),
             mods  = list('data/visit-five/outcomes.csv',
                          'data/visit-five/adjusted.csv'),
             filts = list('data/visit-five/filters.csv'))

adj <- readr::read_csv('data/visit-five/adjusted.csv')[[1]][-1]

visit.data <- get.data(fifth.visit.study, soma.data.fifth, adj)

num.vars <- numeric.id(visit.data, adj)
visit.data.scaled <- get.scaled(visit.data, num.vars, labels$term)
}
}
