% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic-net.R
\name{get.score}
\alias{get.score}
\title{LASSO Score Constructor}
\usage{
get.score(data, coefs, adjust, time, outc)
}
\arguments{
\item{data}{The tibble containing the merged proteomic and study data.}

\item{coefs}{The LASSO coefficients output from \code{\link[=elastic.net]{elastic.net()}}.}

\item{adjust}{The vector of adjustment variables}

\item{time}{The character vector of the Survival outcome time to event
variable}

\item{outc}{The character vector of the Survival outcome event indicator}
}
\value{
\describe{ \item{cox}{A Cox Model of the LASSO Score}
\item{score}{The actual double vector constructed score} }
}
\description{
This function uses the LASSO Model retained coefficients to recreate the
linear predictor. This predictor is then adjusted and regressed against the
outcome. This predictor is only based on the protein values, thus adjustment
variables must be included.
}
\examples{
\dontrun{

fifth.score <- get.score(fifth.visit, lasso$coefs, adjust,
                         "fuptime", "hfdiag")


}
}
