% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{get.soma}
\alias{get.soma}
\title{Load Raw SomaLogic Proteomic Data From Text File}
\usage{
get.soma(path, labels = NULL, id = "SampleId")
}
\arguments{
\item{path}{File path to the SomaLogic text file with protein data.}

\item{labels}{Output of \code{\link[=get.labels]{get.labels()}} function protein mapping dictionary}

\item{id}{Variable name of Study Identifier in SomaLogic Text File used to
map study subjects to clinical data.}
}
\value{
Tibble containing raw proteomic data for all measured subjects.
Tibble will be labelled if labelling dictionary is provided. Sample ID to
map to subject study id is stored in the \code{id} variable.
}
\description{
This is the function to read and process raw text file proteomics data from
SomaLogic. Data can provided with a label to preserve mapping from SeqId to
UniProt Full Name. Study ID is preserved in the ID data column.
}
\examples{
\dontrun{
labels <- get.labels('data/proteinMapping.csv', sid,
                     uniprot.full.name,
                     entrezgenesymbol,
                     flag2 == 0)

soma.data.fifth <- get.soma('data/somaV5.txt', labels = labels)
soma.data.third <- get.soma('data/somaV3.txt', labels = labels)
}
}
