% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{get.visit}
\alias{get.visit}
\title{Get Master Visit Data File with Proteomics}
\usage{
get.visit(soma.data, path.mods, master)
}
\arguments{
\item{soma.data}{SomaLogic Proteomic Raw Data from \code{\link[=get.soma]{get.soma()}}}

\item{path.mods}{Char vector path to folder structure containing new
adjustment variables, filtering conditions, and outcome variables. This
folder must contain csv files formatted as detailed in \code{\link[=get.adjust]{get.adjust()}}. This
function expects that adjustment variables will be in \code{adjusted.csv},
outcome variables in \code{outcomes.csv} and filters in \code{filters.csv}.}

\item{master}{Tibble of master data file with study data. Output of
\code{\link[haven:read_dta]{haven::read_dta()}} function.}
}
\value{
Tibble of study data merged with Proteomics Data, labelled,
and filtered.
}
\description{
This is the outward facing wrapper function to process and load data for
analysis. This function will read SomaLogic and Study Data, Impute missing
adjusting covariates, filter data for exclusions
}
\examples{
\dontrun{
labels <- get.labels('data/proteinMapping.csv', sid,
                     uniprot.full.name,
                     entrezgenesymbol,
                     flag2 == 0)
aric.master <- haven::read_dta('data/ARICmaster_121820.dta')
soma.data <- get.soma('data/somaV5.txt', labels)

fifth.visit <- get.visit(soma.data = soma.data,
                         path.mods = 'data/visit-five',
                         labels = labels,
                         master = aric.master))
}
}
