% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{numeric.id}
\alias{numeric.id}
\title{Get Numeric Variables (for scaling)}
\usage{
numeric.id(data, vars)
}
\arguments{
\item{data}{Tibble of data containing variables to be considered}

\item{vars}{Vector of all variables to be considered}
}
\value{
Vector containing column names of numeric variables
}
\description{
Numeric Values will likely need to be scaled for survival analysis. This
function identifies which variables are numeric, separating them from
categorical factor variables. used in \code{\link[=get.visit]{get.visit()}} and \code{\link[=add.data]{add.data()}}
}
\examples{
\dontrun{
new.vars <- readr::read_csv('data/visit-five/echovars.csv')[[1]]
the.data <- haven::read_dta('data/ARICmaster_121820.dta')
num.vars <- numeric.id(vars = new.vars, data = the.data)
}
}
