% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-array.R
\name{terms.to.keep}
\alias{terms.to.keep}
\title{Filtering Function to Identify which Proteins are retained}
\usage{
\method{terms}{to.keep}(data, .pval, .src, .terms = NULL)
}
\arguments{
\item{data}{The Univariate (and Validation) Tibble with Regression Results.
Output of \code{\link[=cox.arry]{cox.arry()}}.}

\item{.pval}{The significance level to examine which proteins are retained.}

\item{.src}{The tibble in which we examine retained proteins. .src = 1 means
the Derivation dataset if the derivation and validation sets are row-bound.}

\item{.terms}{Vector of terms to limit examination of significance to. Not
necessary to be specified.}
}
\value{
\describe{
\item{kept}{A vector containing all Proteins that were retained by the filter}
\item{data}{A Tibble with the Hazard Ratio and Significance of Retained Proteins}
}
}
\description{
This function calculates which proteins are retained at specified
significance intervals. For example, when the univariate results are
provided, this function returns all proteins that are bonferroni significant.
}
\details{
This function can also be used to check which proteins are retained in
validation.
}
\examples{
\dontrun{
plot.data <-
  dplyr::bind_rows(univariate.results,
                   validation.results) \%>\%
  dplyr::mutate(Name = labels[term, "name"])

bonferroni <-
  terms.to.keep(plot.data,
                .pval = 0.05/length(proteins), 1)
falsediscr <-
  terms.to.keep(plot.data,
                .pval = 0.05, 1)

retained <-
  terms.to.keep(plot.data,
                .pval = 0.05, 2,
                .terms = bonferroni$kept)

}
}
