% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-array.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano Plot Generator To Summarize Univariate Protein Significance}
\usage{
volcanoPlot(data, .pval1, .pval2, suffix, ...)
}
\arguments{
\item{data}{The plot data (with vars \code{hazr} and \code{pval}) to create
volcano plot with.}

\item{.pval1}{The significance indicator for scaling 0.05 for the derivation
dataset. The number of comparisons must be provided (i.e. 4877 for
bonferroni or 1 for FDR).}

\item{.pval2}{The significance indicator for scaling 0.05 for the validation
dataset. The number of comparisons must be provided (i.e. 4877 for
bonferroni or 1 for FDR).}

\item{suffix}{The Suffix that is used on the data join to separate the
Derivation and Validation Sets.}

\item{...}{Additional Parameters (such as title) to be passed to
\code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}}
}
\value{
A Volcano Plot, comparing Hazard Ratio against -log10(pvalue)
}
\description{
This function takes in plot data and uses it to construct a Volcano Plot to
visualize significance. The legend is dependent on the derivation and
validation sets. As currently implemented, this function must have both
datasets bound.
}
\examples{
\dontrun{

plot.data.visit.five <-
  dplyr::left_join(univariate.results,
                   validation.results,
                   by = "term",
                   suffix = c(".V5", ".V3"))

  volcanoPlot(plot.data.visit.five,
              length(proteins),
              length(bonferroni$kept),
              suffix = c(".V5", ".V3"))

}
}
